refactor : RetrofitClient 싱글톤으로 변경
