+++ domain/src/main/java/woowacourse/shopping/configure/ApplicationConfigure.kt
@@ -1,0 +1,5 @@
+package woowacourse.shopping.configure
+
+object ApplicationConfigure {
+    lateinit var BASE_URL: String
+}
